Added logger to debug, moved test files

+++ arthritis_proj/tests/test_data_proc.py
@@ -10,24 +10,35 @@ import unittest
 from contextlib import contextmanager
 from io import StringIO
 import os.path
-
 import numpy as np
+import logging
+from arthritis_proj import main, DEFAULT_DATA_FILE_LOC, data_analysis
 
-from arthritis_proj import main, DEFAULT_DATA_FILE_LOC, data_analysis, DATA_DIR
+TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), 'data_proc')
+
+logging.basicConfig(level=logging.DEBUG)
+logger = logging.getLogger(__name__)
+DISABLE_REMOVE = logger.isEnabledFor(logging.DEBUG)
 
 
 class TestMain(unittest.TestCase):
     def setUp(self):
         # Create a temporary directory
         self.test_dir = tempfile.mkdtemp()
+        if logger.isEnabledFor(logging.DEBUG):
+            print("Writing test output to: {}".format(self.test_dir))
         os.chdir(self.test_dir)
 
     def tearDown(self):
         # Remove the directory after the test
         os.chdir(os.path.dirname(__file__))
-        shutil.rmtree(self.test_dir)
+        if logger.isEnabledFor(logging.DEBUG):
+            print("Not removing temporary files from: {}".format(self.test_dir))
+        else:
+            shutil.rmtree(self.test_dir)
 
     def testNoArgs(self):
+        # Checks that runs with defaults and that files are created
         test_input = []
         main(test_input)
         with capture_stdout(main, test_input) as output:
@@ -47,23 +58,30 @@ class TestDataAnalysis(unittest.TestCase):
     def setUp(self):
         # Create a temporary directory
         self.test_dir = tempfile.mkdtemp()
+        if logger.isEnabledFor(logging.DEBUG):
+            print("Writing test output to: {}".format(self.test_dir))
         os.chdir(self.test_dir)
 
     def tearDown(self):
         # Remove the directory after the test
         os.chdir(os.path.dirname(__file__))
-        shutil.rmtree(self.test_dir)
+        if logger.isEnabledFor(logging.DEBUG):
+            print("Not removing temporary files from: {}".format(self.test_dir))
+        else:
+            shutil.rmtree(self.test_dir)
 
     def testSampleData(self):
+        # Tests csv output from default input file
         csv_data = np.loadtxt(fname=DEFAULT_DATA_FILE_LOC, delimiter=',')
         analysis_results = data_analysis(csv_data)
-        expected_results = np.loadtxt(fname=os.path.join(DATA_DIR, "sample_data_results.csv"), delimiter=',')
+        expected_results = np.loadtxt(fname=os.path.join(TEST_DATA_DIR, "sample_data_results.csv"), delimiter=',')
         self.assertTrue(np.allclose(expected_results, analysis_results))
 
     def testSampleData2(self):
-        csv_data = np.loadtxt(fname=os.path.join(DATA_DIR, "sample_data2.csv"), delimiter=',')
+        # A second check, with slightly different values
+        csv_data = np.loadtxt(fname=os.path.join(TEST_DATA_DIR, "sample_data2.csv"), delimiter=',')
         analysis_results = data_analysis(csv_data)
-        expected_results = np.loadtxt(fname=os.path.join(DATA_DIR, "sample_data2_results.csv"), delimiter=',')
+        expected_results = np.loadtxt(fname=os.path.join(TEST_DATA_DIR, "sample_data2_results.csv"), delimiter=',')
         self.assertTrue(np.allclose(expected_results, analysis_results))
 
 
@@ -79,6 +97,7 @@ def capture_stdout(command, *args, **kwargs):
     sys.stdout.seek(0)
     yield sys.stdout.read()
     sys.stdout = out
+
 
 @contextmanager
 def capture_stderr(command, *args, **kwargs):
